Move to Command directory
